fix: moving  getIconForLink to utils packages
