+++ packages/utils/src/index.ts
@@ -12,3 +12,4 @@ export * from "./string";
 export * from "./theme";
 export * from "./workspace";
 export * from "./work-item";
+export * from "./get-icon-for-link";
